refactor(focus)!: type text emphasis as enum class FontStyle
Change the emphasis parameter on GlyphProvider::glyphForCodepoint /
metricsForCodepoint / supportsEmphasis, the Font wrappers, and
TextLayout::measureLineWrap's initialEmphasis from uint8_t to a new flags enum
focus::FontStyle { Regular=0, Italic=1<<0, Bold=1<<1 } (text/FontStyle.hpp, with
|, & and ~ operators and a contains() helper). StyledGlyphProvider's variant
indexing and FallbackGlyphProvider's bitmask reads use static_cast<uint8_t> /
contains().

Document the reserved in-band control bytes: TextControlCode names SO/SI/BS/FF,
and applyEmphasisShift folds an SO/SI byte into a running depth, replacing six
duplicated scanners across TextLayout and CanvasView with one helper.
CanvasView::emphasisDepth becomes uint8_t so the shared helper serves its render
path directly. TextLayout.hpp's doc comments are normalized to /** blocks.

No behavioral change.

+++ include/views/CanvasView.hpp
@@ -174,8 +174,9 @@ protected:
     bool lastWasNewline = false;  // Tracks consecutive newlines for paragraph detection
     TextAlignment textAlignment = TextAlignment::Left;
 
-    // Emphasis state (SO/SI control codes): 0=normal, 1=italic, 2=bold, 3=bold+italic
-    int emphasisDepth = 0;
+    // Emphasis depth tracked from SO/SI control codes (0-3). static_cast to
+    // FontStyle for glyph queries (0=regular, 1=italic, 2=bold, 3=bold+italic).
+    uint8_t emphasisDepth = 0;
 
     // Word position tracking (set by setWordMapOutput, used during text rendering)
     std::vector<WordPosition> *wordMapOutput = nullptr;
